+++ scripts/components/AddTask/AddTask.styles.js
@@ -3,7 +3,8 @@ import styled from 'styled-components';
 import { Input as FormInput } from '@components/Forms';
 
 export const Container = styled.form`
-    padding: 1rem 1rem 0;
+    padding: 0 1rem;
+    margin-bottom: 0.8rem;
     display: ${({ hidden }) => (hidden ? 'none' : 'block')};
 `;