+++ scripts/components/Icon/Icon.tsx
@@ -2,13 +2,22 @@ import React from 'react';
 
 import { IconContainer } from './Icon.styles.js';
 
-function Icon({ children, onClick, icon, className, ...props }) {
+interface Props {
+    children?: React.ReactNode;
+    onClick?: (e?: React.MouseEvent<HTMLDivElement>) => void;
+    icon: any;
+    className?: string;
+    color?: string;
+}
+
+function Icon({ children, onClick, icon, className, color, ...props }: Props) {
     return (
         <IconContainer
             {...props}
+            color={color}
             as={onClick ? 'button' : 'div'}
             className={className}
-            onClick={onClick ? e => onClick(e) : null}
+            onClick={onClick}
         >
             {React.createElement(icon, {
                 title: children,