+++ scripts/containers/Body/Body.styles.js
@@ -42,6 +42,7 @@ export const Container = styled.div`
     grid-row: 4;
     grid-column: 1;
     overflow: hidden;
+    padding-top: 1rem;
     background: ${({ theme }) => theme.colors.body.background};
     @media ${QUERIES.medium} {
         grid-row: 2;